// ir-gui/src/app/core/enums/sidebar-menu.enum.ts
export enum MenuEnum {
  DASHBOARD = "dashboard",
  MASTER = "master",
  SETTING = "setting",
  AUDIT = "audit",
  LOGOUT = "logout",
  PRODUCT = "product",
  RATING = "rating",
  CUSTOMER = "pre_cust",
  PRE_CUST = "pre_cust",
}

export enum ChildMenuEnum {
  PRE_CUSTS_LIST = "pre_custs_list",
  CUSTOMER_STATUS = "pre_cust_caf_cust_status",
  CUSTOMERLIST = "pre_cust_list",
  RATINGCOUNTRY = "rating_country",
  RATINGPREFIX = "rating_prefix",
  COUNTRY = "country",
  CITY = "city",
  STATE = "state",
  ROLE = "role",
  PINCODE = "pincode",
  SERVICE_AREA = "service_area",
  AREA = "area",
  DEPARTMENT = "department",
  INVESTMENT_CODE = "investment_code",
  BANK = "bank",
  BUSINESS_UNIT = "business_unit",
  SUB_BUSINESS_UNIT = "sub_business_unit",
  BRANCH = "branch",
  REGION = "region",
  BUSINESS_VERTICALS = "business_verticals",
  SUB_BUSINESS_VERTICAL = "sub_business_vertical",
  LOCATION_MASTER = "location_master",
  STAFF = "staff",
  MY_PROFILE = "my_profile",
  SYSTEM_CONFIGURATION = "system_configuration",
  NOTIFICATION_TEMPLATE = "template",
  PAYMENT_GATEWAY_CONFIGURATION = "payment_gateway_configuration",
  MVNO_MANAGEMENT = "mvno_management",
  AUDIT_LOG = "audit_log",
  REPORTED_PROBLEM = "reported_problem",
  PARTNERS = "partners",
  ZONE = "zone",
  TAX_CONFIG = "tax_config",
  TAP_SUMMARY = "tap_summary",
}